Do not infer attributes for virtual methods
Add a flag, `include-virtual-methods`, to include inference of virtual
methods.

The reason this is not enabled by default is because DLP is currently
only looking at the current file when inferring attributes. That means
an overridden method might not have the same inferred attributes as
the parent method. Also, not all subclasses might be known.