Print unit test summary to stdout
PowerShell considers anything printed to stderr as an error. The
default unit test runner implemented by druntime will print a summary
to stderr. This will cause the CI pipeline on AppVeyor to fail.

Since there's no API to change where the summary is printed, it's
necessary to reimplement the unit test runner.